- Adapter classes were moved, model classes were put into model directory.
