--- app/src/main/java/com/netmera/netmerafintech/data/Card.kt
@@ -1,5 +1,0 @@
-package com.netmera.netmerafintech.data
-
-data class Card(
-    var cardId: Int = 0,
-)
