+++ app/src/main/java/com/netmera/netmerafintech/data/Enums.kt
@@ -1,0 +1,7 @@
+package com.netmera.netmerafintech.data
+
+enum class CardType(val value: Int) {
+    CARD_BLACK(0),
+    CARD_BLUE(1),
+    CARD_YELLOW(2),
+}
