+++ app/src/main/java/com/netmera/netmerafintech/data/model/Card.kt
@@ -1,0 +1,5 @@
+package com.netmera.netmerafintech.data.model
+
+data class Card(
+    var cardId: Int = 0,
+)
