--- app/src/main/java/com/netmera/netmerafintech/ui/all_pages/AllPagesViewModel.kt
@@ -1,7 +1,0 @@
-package com.netmera.netmerafintech.ui.all_pages
-
-import androidx.lifecycle.ViewModel
-
-class AllPagesViewModel : ViewModel() {
-    fun test(){}
-}
