--- app/src/main/java/com/netmera/netmerafintech/utils/Constants.kt
@@ -1,5 +1,0 @@
-package com.netmera.netmerafintech.utils
-
-const val CARD_BLACK = 0
-const val CARD_BLUE = 1
-const val CARD_YELLOW = 2
